requirements.txt
kiteconnect
pandas
numpy
requests
psycopg2-binary
python-dotenv
openai
google-generativeai